V3 client factory
