+++ internal/extsvc/github/client.go
@@ -1,7 +1,6 @@
 package github
 
 import (
-	"context"
 	"fmt"
 	"net/http"
 	"net/url"
@@ -9,80 +8,8 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/sourcegraph/sourcegraph/internal/extsvc/auth"
-	"github.com/sourcegraph/sourcegraph/internal/httpcli"
 	"github.com/sourcegraph/sourcegraph/internal/ratelimit"
 )
-
-// NewClient creates a new GitHub API client with an optional default
-// authenticator.
-//
-// apiURL must point to the base URL of the GitHub API. See the docstring for
-// Client.apiURL.
-func NewClient(apiURL *url.URL, a auth.Authenticator, cli httpcli.Doer) *V3Client {
-	apiURL = canonicalizedURL(apiURL)
-	if gitHubDisable {
-		cli = disabledClient{}
-	}
-	if cli == nil {
-		cli = httpcli.ExternalDoer()
-	}
-
-	cli = requestCounter.Doer(cli, func(u *url.URL) string {
-		// The first component of the Path mostly maps to the type of API
-		// request we are making. See `curl https://api.github.com` for the
-		// exact mapping
-		var category string
-		if parts := strings.SplitN(u.Path, "/", 3); len(parts) > 1 {
-			category = parts[1]
-		}
-		return category
-	})
-
-	rl := ratelimit.DefaultRegistry.Get(apiURL.String())
-
-	return &V3Client{
-		apiURL:           apiURL,
-		githubDotCom:     urlIsGitHubDotCom(apiURL),
-		auth:             a,
-		httpClient:       cli,
-		rateLimitMonitor: &ratelimit.Monitor{HeaderPrefix: "X-"},
-		repoCache:        newRepoCache(apiURL, a),
-		rateLimit:        rl,
-	}
-}
-
-// WithAuthenticator returns a new Client that uses the same configuration as
-// the current Client, except authenticated as the GitHub user with the given
-// authenticator instance (most likely a token).
-func (c *V3Client) WithAuthenticator(a auth.Authenticator) *V3Client {
-	return NewClient(c.apiURL, a, c.httpClient)
-}
-
-func (c *V3Client) requestGet(ctx context.Context, requestURI string, result interface{}) error {
-	req, err := http.NewRequest("GET", requestURI, nil)
-	if err != nil {
-		return err
-	}
-
-	// Include node_id (GraphQL ID) in response. See
-	// https://developer.github.com/changes/2017-12-19-graphql-node-id/.
-	//
-	// Enable the repository topics API. See
-	// https://developer.github.com/v3/repos/#list-all-topics-for-a-repository
-	req.Header.Add("Accept", "application/vnd.github.jean-grey-preview+json,application/vnd.github.mercy-preview+json")
-
-	// Enable the GitHub App API. See
-	// https://developer.github.com/v3/apps/installations/#list-repositories
-	req.Header.Add("Accept", "application/vnd.github.machine-man-preview+json")
-
-	err = c.rateLimit.Wait(ctx)
-	if err != nil {
-		return errors.Wrap(err, "rate limit")
-	}
-
-	return doRequest(ctx, c.apiURL, c.auth, c.rateLimitMonitor, c.httpClient, req, result)
-}
 
 // RateLimitMonitor exposes the rate limit monitor.
 func (c *V3Client) RateLimitMonitor() *ratelimit.Monitor {
